requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
tqdm==4.66.5
lxml==5.2.1
python-dotenv==1.0.1
openai==1.55.0
